Login Customer done. Added additional interface for vendor
