+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".ui.login_customer.LoginCustomerActivity"/>
+        <activity android:name=".ui.login_vendor.LoginVendorActivity"/>
     </application>
 
 </manifest>
