+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,7 @@
 <resources>
     <string name="app_name">Eat@NUS</string>
-    <string name="login">Login </string>
+    <string name="login_customer">Login as Customer </string>
+    <string name="login_vendor">Login as Vendor</string>
+    <string name="login"> Login </string>
     <string name="login_name_placeholder">Enter your name</string>
 </resources>
